app/requirements.txt
uvicorn
sqlalchemy
psycopg2-binary
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
bcrypt
requests
asyncpg 
passlib
PyJWT